--- tmp_test_chunk.py
@@ -1,2 +1,0 @@
-from toksum.simple_chunker import SimpleChunker
-print("Import successful")
