--- toksum/chunker.py
@@ -1,280 +1,0 @@
-"""
-Smart text chunking functionality for toksum.
-
-This module provides intelligent text chunking that respects token limits
-while maintaining semantic boundaries like sentences, paragraphs, and code blocks.
-"""
-
-import re
-from typing import List, Optional
-from .core import count_tokens
-
-
-class SmartChunker:
-    """
-    Intelligent text chunker that splits text while respecting token limits.
-    
-    The chunker attempts to maintain semantic boundaries by splitting on:
-    1. Sentences (for general text)
-    2. Paragraphs (for longer documents)
-    3. Code blocks (for source code)
-    
-    Args:
-        model: The model name to use for token counting
-        max_tokens: Maximum tokens per chunk
-    """
-    
-    def __init__(self, model: str, max_tokens: int):
-        """Initialize the SmartChunker."""
-        if max_tokens <= 0:
-            raise ValueError("max_tokens must be positive")
-        
-        self.model = model
-        self.max_tokens = max_tokens
-    
-    def chunk_by_sentences(self, text: str) -> List[str]:
-        """
-        Chunk text by sentences while respecting token limits.
-        
-        Args:
-            text: The text to chunk
-            
-        Returns:
-            List of text chunks, each under the token limit
-        """
-        if not text.strip():
-            return []
-        
-        # Split into sentences using regex
-        sentences = re.split(r'(?<=[.!?])\s+', text.strip())
-        if not sentences:
-            return []
-        
-        chunks: List[str] = []
-        current_chunk: str = ""
-        
-        for sentence in sentences:
-            # Try adding this sentence to current chunk
-            test_chunk = current_chunk
-            if current_chunk:
-                test_chunk += " " + sentence
-            else:
-                test_chunk = sentence
-            
-            # Check if it fits within token limit
-            token_count = count_tokens(test_chunk, self.model)
-            
-            if token_count <= self.max_tokens:
-                current_chunk = test_chunk
-            else:
-                # Current chunk is full, start a new one
-                if current_chunk:
-                    chunks.append(current_chunk)
-                # Check if single sentence fits
-                single_token_count = count_tokens(sentence, self.model)
-                current_chunk = sentence
-        
-        # Add the last chunk if it exists
-        if current_chunk:
-            chunks.append(current_chunk)
-        
-        return chunks
-    
-    def chunk_by_paragraphs(self, text: str) -> List[str]:
-        """
-        Chunk text by paragraphs while respecting token limits.
-        
-        Args:
-            text: The text to chunk
-            
-        Returns:
-            List of text chunks, each under the token limit
-        """
-        if not text.strip():
-            return []
-        
-        # Split into paragraphs
-        paragraphs = re.split(r'\n\s*\n', text.strip())
-        if not paragraphs:
-            return []
-        
-        chunks: List[str] = []
-        current_chunk: str = ""
-        
-        for paragraph in paragraphs:
-            # Try adding this paragraph to current chunk
-            test_chunk = current_chunk
-            if current_chunk:
-                test_chunk += "\n\n" + paragraph
-            else:
-                test_chunk = paragraph
-            
-            # Check if it fits within token limit
-            token_count = count_tokens(test_chunk, self.model)
-            
-            if token_count <= self.max_tokens:
-                current_chunk = test_chunk
-            else:
-                # Current chunk is full, start a new one
-                if current_chunk:
-                    chunks.append(current_chunk)
-                # Check if single paragraph fits, if not it goes alone
-                single_token_count = count_tokens(paragraph, self.model)
-                current_chunk = paragraph
-        
-        # Add the last chunk if it exists
-        if current_chunk:
-            chunks.append(current_chunk)
-        
-        return chunks
-    
-    def chunk_code(self, code: str, language: str = "python") -> List[str]:
-        """
-        Chunk code while respecting token limits and code structure.
-        
-        For Python code, attempts to split on function/class boundaries.
-        For other languages, falls back to paragraph chunking.
-        
-        Args:
-            code: The code to chunk
-            language: Programming language (currently only "python" is specially handled)
-            
-        Returns:
-            List of code chunks, each under the token limit
-        """
-        if not code.strip():
-            return []
-        
-        if language.lower() == "python":
-            return self._chunk_python_code(code)
-        else:
-            # Fallback to paragraph chunking for other languages
-            return self.chunk_by_paragraphs(code)
-    
-    def _chunk_python_code(self, code: str) -> List[str]:
-        """
-        Chunk Python code by function and class definitions.
-        
-        Args:
-            code: Python code to chunk
-            
-        Returns:
-            List of code chunks
-        """
-        # Split code into logical blocks (functions, classes, and other code)
-        blocks = self._split_python_into_blocks(code)
-        
-        if not blocks:
-            return []
-        
-        chunks: List[str] = []
-        current_chunk: str = ""
-        
-        for block in blocks:
-            if not current_chunk:
-                # First block - check its token count and start chunk
-                token_count = count_tokens(block, self.model)
-                current_chunk = block
-            else:
-                # Check if current block alone fits within limits
-                block_count = count_tokens(block, self.model)
-                
-                # Try combining with current chunk
-                test_chunk = current_chunk + "\n\n" + block
-                combined_count = count_tokens(test_chunk, self.model)
-                
-                # Special handling: if we have both function and class, 
-                # and the combined text is long, it might exceed limits
-                # even if the mock returns a low value due to pattern matching
-                has_function = "def " in current_chunk
-                has_class = "class " in current_chunk
-                new_has_function = "def " in block
-                new_has_class = "class " in block
-                
-                # If combining function and class, and text is long, split them
-                if ((has_function and new_has_class) or (has_class and new_has_function)) and len(test_chunk) > 100:
-                    chunks.append(current_chunk)
-                    current_chunk = block
-                elif combined_count <= self.max_tokens:
-                    current_chunk = test_chunk
-                else:
-                    # Doesn't fit, finalize current chunk and start new one
-                    chunks.append(current_chunk)
-                    current_chunk = block
-        
-        # Add the final chunk
-        if current_chunk:
-            chunks.append(current_chunk)
-        
-        return chunks
-    
-    def _split_python_into_blocks(self, code: str) -> List[str]:
-        """
-        Split Python code into logical blocks (functions, classes, other code).
-        
-        Args:
-            code: Python code to split
-            
-        Returns:
-            List of code blocks
-        """
-        lines = code.strip().split('\n')
-        blocks: List[str] = []
-        current_block_lines: List[str] = []
-        
-        i: int = 0
-        while i < len(lines):
-            line = lines[i]
-            
-            # Check if this line starts a new top-level function or class
-            if re.match(r'^(def |class )', line):
-                # If we have accumulated non-function/class lines, save them as a block
-                if current_block_lines:
-                    block_content = '\n'.join(current_block_lines).strip()
-                    if block_content:
-                        blocks.append(block_content)
-                    current_block_lines = []
-                
-                # Collect the entire function or class definition
-                function_lines: List[str] = [line]
-                i += 1
-                
-                # Get the indentation level of the function/class
-                base_indent: int = len(line) - len(line.lstrip())
-                
-                # Collect all lines that belong to this function/class
-                while i < len(lines):
-                    next_line = lines[i]
-                    
-                    # If it's an empty line, include it
-                    if not next_line.strip():
-                        function_lines.append(next_line)
-                        i += 1
-                        continue
-                    
-                    # If it's indented more than the function/class, it belongs to it
-                    next_indent: int = len(next_line) - len(next_line.lstrip())
-                    if next_indent > base_indent:
-                        function_lines.append(next_line)
-                        i += 1
-                        continue
-                    
-                    # If we hit another top-level definition or unindented code, stop
-                    break
-                
-                # Add the complete function/class as a block
-                block_content = '\n'.join(function_lines).strip()
-                if block_content:
-                    blocks.append(block_content)
-                continue
-            else:
-                current_block_lines.append(line)
-                i += 1
-        
-        # Add any remaining lines as a block
-        if current_block_lines:
-            block_content = '\n'.join(current_block_lines).strip()
-            if block_content:
-                blocks.append(block_content)
-        
-        return blocks
